refactoring publipostage
